+++ source/Core/Database/Adapter/Doctrine/Database.php
@@ -18,6 +18,8 @@ use OxidEsales\Eshop\Core\Database\Adapter\DatabaseInterface;
 use OxidEsales\Eshop\Core\Exception\DatabaseConnectionException;
 use OxidEsales\Eshop\Core\Exception\DatabaseErrorException;
 use OxidEsales\Eshop\Core\Exception\StandardException;
+use OxidEsales\EshopCommunity\Internal\Framework\Database\Logger\DatabaseLoggerFactoryInterface;
+use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
 use PDO;
 
 /**
@@ -94,12 +96,6 @@ class Database implements DatabaseInterface
     public function connect()
     {
         $connection = null;
-
-        /**
-         * @todo we need a SQLLogger that logs to a (CSV?) file, as we probably do not want to log into the database.
-         *
-         * $configuration->setSQLLogger(new EchoSQLLogger());
-         */
 
         try {
             $connection = $this->getConnectionFromDriverManager();
@@ -1382,6 +1378,11 @@ class Database implements DatabaseInterface
     {
         $configuration = new Configuration();
         $connectionParameters = $this->getConnectionParameters();
+
+        $container = ContainerFactory::getInstance()->getContainer();
+        $configuration->setSQLLogger(
+            $container->get(DatabaseLoggerFactoryInterface::class)->getDatabaseLogger()
+        );
 
         return DriverManager::getConnection($connectionParameters, $configuration);
     }