+++ source/Internal/Framework/Logger/LoggerServiceFactory.php
@@ -62,7 +62,8 @@ class LoggerServiceFactory
     {
         return new MonologConfiguration(
             'OXID Logger',
-            $this->context
+            $this->context->getLogFilePath(),
+            $this->context->getLogLevel()
         );
     }