+++ source/Internal/Transition/Utility/Context.php
@@ -6,12 +6,14 @@
 
 namespace OxidEsales\EshopCommunity\Internal\Transition\Utility;
 
-use OxidEsales\Eshop\Core\Config;
 use OxidEsales\EshopCommunity\Core\DatabaseProvider;
-use OxidEsales\EshopCommunity\Internal\Transition\Utility\BasicContext;
 use PDO;
 use Psr\Log\LogLevel;
 use Webmozart\PathUtil\Path;
+use OxidEsales\Facts\Config\ConfigFile as FactsConfigFile;
+use OxidEsales\Eshop\Core\Registry;
+use OxidEsales\Eshop\Core\Config;
+use OxidEsales\EshopCommunity\Internal\Transition\Utility\Exception\AdminUserNotFoundException;
 
 /**
  * @internal
@@ -19,17 +21,9 @@ use Webmozart\PathUtil\Path;
 class Context extends BasicContext implements ContextInterface
 {
     /**
-     * @var Config
+     * @var FactsConfigFile
      */
-    private $config;
-
-    /**
-     * @param Config $config
-     */
-    public function __construct(Config $config)
-    {
-        $this->config = $config;
-    }
+    private $factsConfigFile;
 
     /**
      * @return string
@@ -44,7 +38,7 @@ class Context extends BasicContext implements ContextInterface
      */
     public function getLogFilePath(): string
     {
-        return Path::join($this->config->getLogsDir(), 'oxideshop.log');
+        return Path::join(Registry::getConfig()->getLogsDir(), 'oxideshop.log');
     }
 
     /**
@@ -62,7 +56,7 @@ class Context extends BasicContext implements ContextInterface
      */
     public function getCurrentShopId(): int
     {
-        return $this->config->getShopId();
+        return Registry::getConfig()->getShopId();
     }
 
     /**
@@ -72,7 +66,7 @@ class Context extends BasicContext implements ContextInterface
     {
         $integerShopIds = [];
 
-        foreach ($this->config->getShopIds() as $shopId) {
+        foreach (Registry::getConfig()->getShopIds() as $shopId) {
             $integerShopIds[] = (int) $shopId;
         }
 
@@ -88,6 +82,60 @@ class Context extends BasicContext implements ContextInterface
     }
 
     /**
+     * @return bool
+     */
+    public function isAdmin(): bool
+    {
+        return $this->isConfigLoaded() ? Registry::getConfig()->isAdmin() : isAdmin();
+    }
+
+    /**
+     * @return bool
+     */
+    public function isEnabledAdminQueryLog(): bool
+    {
+        return (bool) $this->getFactsConfigFile()->getVar('blLogChangesInAdmin');
+    }
+
+    /**
+     * @return string
+     */
+    public function getAdminLogFilePath(): string
+    {
+        return Path::join($this->getSourcePath(), 'log', 'oxadmin.log');
+    }
+
+    /**
+     * We need to be careful when trying to fetch config parameters in this place as the
+     * shop might still be bootstrapping.
+     * The config must be already initialized before we can safely call Config::getConfigParam().
+     *
+     * @return array
+     */
+    public function getSkipLogTags(): array
+    {
+        $skipLogTags = [];
+        if ($this->isConfigLoaded()) {
+            $skipLogTags = Registry::getConfig()->getConfigParam('aLogSkipTags');
+        }
+
+        return (array) $skipLogTags;
+    }
+
+    /**
+     * @return string
+     */
+    public function getAdminUserId(): string
+    {
+        $adminUserId = (string) Registry::getSession()->getVariable('auth');
+        if (empty($adminUserId)) {
+            throw new AdminUserNotFoundException();
+        }
+
+        return $adminUserId;
+    }
+
+    /**
      * @param string $name
      * @param null   $default
      *
@@ -95,8 +143,28 @@ class Context extends BasicContext implements ContextInterface
      */
     private function getConfigParameter($name, $default = null)
     {
-        $value = $this->config->getConfigParam($name, $default);
+        $value = Registry::getConfig()->getConfigParam($name, $default);
         DatabaseProvider::getDb()->setFetchMode(PDO::FETCH_ASSOC);
         return $value;
+    }
+
+    /**
+     * @return FactsConfigFile
+     */
+    private function getFactsConfigFile(): FactsConfigFile
+    {
+        if (!is_a($this->factsConfigFile, FactsConfigFile::class)) {
+            $this->factsConfigFile = new FactsConfigFile();
+        }
+
+        return $this->factsConfigFile;
+    }
+
+    /**
+     * @return bool
+     */
+    private function isConfigLoaded(): bool
+    {
+        return Registry::instanceExists(Config::class);
     }
 }