+++ source/Internal/Transition/Utility/ContextInterface.php
@@ -6,7 +6,7 @@
 
 namespace OxidEsales\EshopCommunity\Internal\Transition\Utility;
 
-use OxidEsales\EshopCommunity\Internal\Transition\Utility\BasicContextInterface;
+use OxidEsales\EshopCommunity\Internal\Transition\Utility\Exception\AdminUserNotFoundException;
 
 /**
  * @internal
@@ -37,4 +37,32 @@ interface ContextInterface extends BasicContextInterface
      * @return string
      */
     public function getConfigurationEncryptionKey(): string;
+
+    /**
+     * @return bool
+     */
+    public function isEnabledAdminQueryLog(): bool;
+
+    /**
+     * @return bool
+     */
+    public function isAdmin(): bool;
+
+    /**
+     * @return string
+     */
+    public function getAdminLogFilePath(): string;
+
+    /**
+     * @return array
+     */
+    public function getSkipLogTags(): array;
+
+    /**
+     * @return string
+     *
+     * @throws AdminUserNotFoundException
+     */
+    public function getAdminUserId(): string;
+
 }